fix(builder): align workspace invite modal existing-user behavior
